requirements.txt
Django==6.0.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
django-environ==0.11.2
psycopg2-binary==2.9.10
Pillow==11.2.1
# Production
gunicorn==21.2.0
whitenoise[brotli]==6.8.2
